nanoOffset needed to be reset
git-svn-id: https://svn1.hosted-projects.com/transparentpolitics/amplafi-flow/amplafi-sworddance/trunk@191 6eaf30a2-012b-4626-ba8e-69dab18af4f2

+++ src/main/java/com/sworddance/scheduling/FakeTimeServer.java
@@ -51,7 +51,7 @@ public interface FakeTimeServer extends TimeServer {
      * @param duration
      * @param units
      */
-    public void runClock(int duration, TimeUnit units);
+    public void runClock(long duration, TimeUnit units);
 
     public long getNanoOffset();
